Optional refresh token in implicit grant
Solution proposal for issues: #98.

This commit includes additional logic to make the generation
of refresh tokens optional in implicit grant flows so that this package
becomes strictly compliant with the OAuth 2.0 specification in section
4.2. To see further information, please refer to issue #98.

To not break backward compatibility, this new option has been included
as an new optional setting flag, `ISSUE_REFRESH_TOKEN_IMPLICIT_GRANT`,
that when set to `False`, it deactivates the refresh token generation in
implicit flows.